Support for delayedSendToQueue
A mix between sendToQueue and delayedPublish. Works by creating/reusing
an exchange with a specific delay which is bound to a queue specific to
the delay. This queue has its' dead letter exchange set to "" (the
default exchange) and when it dead-letters a message with routing key X
it will be sent to the queue X (this is the behaviour when publishing a
message on the default exhange).